+++ packages/xrpc/lib/src/xrpc/exception/internal_server_error_exception.dart
