+++ packages/xrpc/lib/src/xrpc/exception/invalid_request_exception.dart
