+++ packages/xrpc/lib/src/xrpc/exception/unauthorized_exception.dart
