+++ packages/xrpc/lib/src/xrpc/exception/xrpc_exception.dart
