+++ packages/xrpc/lib/src/xrpc/exception/xrpc_not_supported_exception.dart
