+++ packages/xrpc/lib/src/xrpc/xrpc_error.dart
