+++ packages/xrpc/lib/xrpc.dart
@@ -8,18 +8,19 @@ export 'package:xrpc/src/types.dart'
 export 'package:xrpc/src/entities/empty_data.dart';
 export 'package:xrpc/src/entities/rate_limit.dart';
 export 'package:xrpc/src/entities/rate_limit_policy.dart';
-export 'package:xrpc/src/exception/internal_server_error_exception.dart';
-export 'package:xrpc/src/exception/invalid_request_exception.dart';
-export 'package:xrpc/src/exception/rate_limit_exceeded_exception.dart';
-export 'package:xrpc/src/exception/unauthorized_exception.dart';
-export 'package:xrpc/src/exception/xrpc_exception.dart';
-export 'package:xrpc/src/exception/xrpc_not_supported_exception.dart';
+export 'package:xrpc/src/xrpc/exception/internal_server_error_exception.dart';
+export 'package:xrpc/src/xrpc/exception/invalid_request_exception.dart';
+export 'package:xrpc/src/xrpc/exception/rate_limit_exceeded_exception.dart';
+export 'package:xrpc/src/xrpc/exception/unauthorized_exception.dart';
+export 'package:xrpc/src/xrpc/exception/xrpc_exception.dart';
+export 'package:xrpc/src/xrpc/exception/xrpc_not_supported_exception.dart';
 export 'package:xrpc/src/http_method.dart';
 export 'package:xrpc/src/http_status.dart';
 export 'package:xrpc/src/protocol.dart';
 export 'package:xrpc/src/serializable.dart';
 export 'package:xrpc/src/subscription.dart';
-export 'package:xrpc/src/xrpc.dart' show query, procedure, upload, subscribe;
-export 'package:xrpc/src/xrpc_error.dart';
-export 'package:xrpc/src/xrpc_request.dart';
-export 'package:xrpc/src/xrpc_response.dart';
+export 'package:xrpc/src/xrpc/xrpc.dart'
+    show query, procedure, upload, subscribe;
+export 'package:xrpc/src/xrpc/xrpc_error.dart';
+export 'package:xrpc/src/xrpc/xrpc_request.dart';
+export 'package:xrpc/src/xrpc/xrpc_response.dart';
