+++ packages/xrpc/test/src/xrpc/xrpc_test.dart
@@ -15,14 +15,14 @@ import 'package:test/test.dart';
 // 🌎 Project imports:
 import 'package:xrpc/src/entities/empty_data.dart';
 import 'package:xrpc/src/entities/rate_limit.dart';
-import 'package:xrpc/src/exception/internal_server_error_exception.dart';
-import 'package:xrpc/src/exception/invalid_request_exception.dart';
-import 'package:xrpc/src/exception/rate_limit_exceeded_exception.dart';
-import 'package:xrpc/src/exception/unauthorized_exception.dart';
-import 'package:xrpc/src/exception/xrpc_not_supported_exception.dart';
 import 'package:xrpc/src/subscription.dart';
-import 'package:xrpc/src/xrpc.dart';
-import 'package:xrpc/src/xrpc_response.dart';
+import 'package:xrpc/src/xrpc/exception/internal_server_error_exception.dart';
+import 'package:xrpc/src/xrpc/exception/invalid_request_exception.dart';
+import 'package:xrpc/src/xrpc/exception/rate_limit_exceeded_exception.dart';
+import 'package:xrpc/src/xrpc/exception/unauthorized_exception.dart';
+import 'package:xrpc/src/xrpc/exception/xrpc_not_supported_exception.dart';
+import 'package:xrpc/src/xrpc/xrpc.dart';
+import 'package:xrpc/src/xrpc/xrpc_response.dart';
 
 void main() {
   group('.checkStatus', () {
